Remade Item Collection
this is pretty big, and also a little spaghetti.

My solution to copying an item into memory is to have an item handler that will store all of the relevant data individually. The player communicates to the item handler to store and remove items, as well as asking it to spawn and destory actors. This way, I am able to assign attributes easily to each object. Not all attributes are assigned right now, but proof of concept works with attributes such as damage, armour, durability, etc. Meshes and Materials are not working (didnt look into how to do it yet, but it must be really simple).