Worked on Bow Shooting
Worked on animating line decals for bow shooting. The basic for this is when you press the shoot button two lines will rotate inwards over a time period. When you release the shoot button before the lines have finished animating the accuracy of the bow will be decreased. i.e., it will fire at a random direction between the two lines.

Still a work in progress. Currently have the lines animated but need to get make the accuracy of the projectile variable. Also need to add constraints to character movement while aiming. None of this needs to be the final way we have bow mechanics, just something I thought of and wanted to program.