Added Bow Powershot
If the bow is released near the apex of the bow line glow, then the bow projectile is spawned at twice the speed. The timing of this can be modified, currently the canPowershot variable is set to true when the timeline from 0 to 1 reaches 0.75. It is set to false when it goes below 0.75 on the reverse timeline.