Moved UI Ability Activation out of Player Manager
Very small update. Moved the triggering of the UI ability activation out of the manager class and into the individual ability class. Did this by overriding the base ability activation method and adding the call to the UI activation function. Just a little bit cleaner and does not rely of using array of abilities in the manager.