added filesearching and CSV processing with custom struct
does not work currently (my hunch is on this one line of code and idk how to fix it...), but the goal is to have my own function that will be able to dynamically read CSV files at runtime based on the name of a character. It currently can find the file and store its contents to a string. None of this code is implemented in-engine, so it sits inert.

ISSUE: Line 102 in ComponentFunctions.cpp:
fstream[charIndex] != *"," doesnt work (for no reason?), even though there are commas in the file.